Revert "Create plugin loader (#1781)"
This reverts commit 23a6952c6d0ec3343bc7704aacd5ec2d89bbb557.
