+++ docs/geedocs/5.3.6/answer/developers.html
@@ -72,7 +72,6 @@
 <ul>
 <li class="toctree-l1"><a class="reference internal" href="developers/addPythonLibToGEPython.html">Add Python Libraries to GEE Python</a></li>
 <li class="toctree-l1"><a class="reference internal" href="developers/googleMapsAPIV3ForGEE.html">Google Maps API v3 for GEE</a></li>
-<li class="toctree-l1"><a class="reference internal" href="developers/creatingFileAccessorPlugin.html">Creating a File Accessor Plugin</a></li>
 </ul>
 </div>
 </div>
