+++ docs/geedocs/5.3.6/answer/developers/googleMapsAPIV3ForGEE.html
@@ -16,7 +16,7 @@
     <script type="text/javascript" src="../../static/bizstyle.js"></script>
     <link rel="index" title="Index" href="../../genindex.html" />
     <link rel="search" title="Search" href="../../search.html" />
-    <link rel="next" title="Creating a File Accessor Plugin" href="creatingFileAccessorPlugin.html" />
+    <link rel="next" title="Troubleshooting and error messages" href="../troubleshooting.html" />
     <link rel="prev" title="Add Python Libraries to GEE Python" href="addPythonLibToGEPython.html" />
     <meta name="viewport" content="width=device-width,initial-scale=1.0">
     <!--[if lt IE 9]>
@@ -30,7 +30,7 @@
           <a href="../../genindex.html" title="General Index"
              accesskey="I">index</a></li>
         <li class="right" >
-          <a href="creatingFileAccessorPlugin.html" title="Creating a File Accessor Plugin"
+          <a href="../troubleshooting.html" title="Troubleshooting and error messages"
              accesskey="N">next</a> |</li>
         <li class="right" >
           <a href="addPythonLibToGEPython.html" title="Add Python Libraries to GEE Python"
@@ -45,8 +45,8 @@
   <p class="topless"><a href="addPythonLibToGEPython.html"
                         title="previous chapter">Add Python Libraries to GEE Python</a></p>
   <h4>Next topic</h4>
-  <p class="topless"><a href="creatingFileAccessorPlugin.html"
-                        title="next chapter">Creating a File Accessor Plugin</a></p>
+  <p class="topless"><a href="../troubleshooting.html"
+                        title="next chapter">Troubleshooting and error messages</a></p>
 <div id="searchbox" style="display: none" role="search">
   <h3>Quick search</h3>
     <div class="searchformwrapper">
@@ -333,7 +333,7 @@ and
           <a href="../../genindex.html" title="General Index"
              >index</a></li>
         <li class="right" >
-          <a href="creatingFileAccessorPlugin.html" title="Creating a File Accessor Plugin"
+          <a href="../troubleshooting.html" title="Troubleshooting and error messages"
              >next</a> |</li>
         <li class="right" >
           <a href="addPythonLibToGEPython.html" title="Add Python Libraries to GEE Python"
