+++ docs/geedocs/5.3.6/answer/troubleshooting.html
@@ -17,7 +17,7 @@
     <link rel="index" title="Index" href="../genindex.html" />
     <link rel="search" title="Search" href="../search.html" />
     <link rel="next" title="Geecheck: test your GEE Server status" href="troubleshooting/geecheckTestServerStatus.html" />
-    <link rel="prev" title="Creating a File Accessor Plugin" href="developers/creatingFileAccessorPlugin.html" />
+    <link rel="prev" title="Google Maps API v3 for GEE" href="developers/googleMapsAPIV3ForGEE.html" />
     <meta name="viewport" content="width=device-width,initial-scale=1.0">
     <!--[if lt IE 9]>
     <script type="text/javascript" src="static/css3-mediaqueries.js"></script>
@@ -33,7 +33,7 @@
           <a href="troubleshooting/geecheckTestServerStatus.html" title="Geecheck: test your GEE Server status"
              accesskey="N">next</a> |</li>
         <li class="right" >
-          <a href="developers/creatingFileAccessorPlugin.html" title="Creating a File Accessor Plugin"
+          <a href="developers/googleMapsAPIV3ForGEE.html" title="Google Maps API v3 for GEE"
              accesskey="P">previous</a> |</li>
         <li class="nav-item nav-item-0"><a href="../index.html">Google Earth Enterprise 5.3.6 documentation</a> &#187;</li> 
       </ul>
@@ -41,8 +41,8 @@
       <div class="sphinxsidebar" role="navigation" aria-label="main navigation">
         <div class="sphinxsidebarwrapper">
   <h4>Previous topic</h4>
-  <p class="topless"><a href="developers/creatingFileAccessorPlugin.html"
-                        title="previous chapter">Creating a File Accessor Plugin</a></p>
+  <p class="topless"><a href="developers/googleMapsAPIV3ForGEE.html"
+                        title="previous chapter">Google Maps API v3 for GEE</a></p>
   <h4>Next topic</h4>
   <p class="topless"><a href="troubleshooting/geecheckTestServerStatus.html"
                         title="next chapter">Geecheck: test your GEE Server status</a></p>
@@ -122,7 +122,7 @@
           <a href="troubleshooting/geecheckTestServerStatus.html" title="Geecheck: test your GEE Server status"
              >next</a> |</li>
         <li class="right" >
-          <a href="developers/creatingFileAccessorPlugin.html" title="Creating a File Accessor Plugin"
+          <a href="developers/googleMapsAPIV3ForGEE.html" title="Google Maps API v3 for GEE"
              >previous</a> |</li>
         <li class="nav-item nav-item-0"><a href="../index.html">Google Earth Enterprise 5.3.6 documentation</a> &#187;</li> 
       </ul>
