+++ docs/geedocs/5.3.6/index.html
@@ -218,7 +218,6 @@ also available in the GEE Server 5.3.6 installation package.</div></blockquote>
 <li class="toctree-l1"><a class="reference internal" href="answer/developers.html">Developers</a><ul>
 <li class="toctree-l2"><a class="reference internal" href="answer/developers/addPythonLibToGEPython.html">Add Python Libraries to GEE Python</a></li>
 <li class="toctree-l2"><a class="reference internal" href="answer/developers/googleMapsAPIV3ForGEE.html">Google Maps API v3 for GEE</a></li>
-<li class="toctree-l2"><a class="reference internal" href="answer/developers/creatingFileAccessorPlugin.html">Creating a File Accessor Plugin</a></li>
 </ul>
 </li>
 <li class="toctree-l1"><a class="reference internal" href="answer/troubleshooting.html">Troubleshooting and error messages</a><ul>
