+++ docs/geedocs/5.3.6/raw_sources/answer/developers.rst.txt
@@ -7,4 +7,3 @@ Developers
 
    developers/addPythonLibToGEPython
    developers/googleMapsAPIV3ForGEE
-   developers/creatingFileAccessorPlugin.rst
