+++ docs/geedocs/docsrc/answer/developers.rst
@@ -7,4 +7,3 @@ Developers
 
    developers/addPythonLibToGEPython
    developers/googleMapsAPIV3ForGEE
-   developers/creatingFileAccessorPlugin.rst
